+++ MORE/MoreEvent.cs
@@ -1,9 +1,14 @@
-﻿namespace Shockah.MORE;
+﻿using Newtonsoft.Json;
+using Newtonsoft.Json.Converters;
 
+namespace Shockah.MORE;
+
+[JsonConverter(typeof(StringEnumConverter))]
 internal enum MoreEvent
 {
 	AbyssalPower,
+	ArtifactSwap,
 	CombatDataCalibration,
 	DraculaDeckTrial,
-	ShipSwap
+	ShipSwap,
 }